refactor deathmatch
